pyproject.toml
[project]
name = "qa-with-followups-agent"
version = "1.0.0"
description = "A2A agent that answers questions and provides follow-up suggestions"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "langchain-google-genai>=2.0.8",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.73",
    "pydantic>=2.0.0",
    "uvicorn>=0.34.0",
    "starlette>=0.46.0",
    "sse-starlette>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
